frozen before reinstall
